services/tokenizer/requirements.txt
flask==0.12.2
pyspellchecker
